+++ R/RcppExports.R
@@ -25,10 +25,11 @@ keyATMvb_call <- function(model) {
 #' @param dfm a dfm input (sparse Matrix)
 #' @param W_raw an object to return
 #' @param vocab a vector of vocabulary
+#' @param show_progress_bar show a progress bar 
 #'
 #' @keywords internal
-read_dfm_cpp <- function(dfm, W_raw, vocab) {
-    .Call(`_keyATM_read_dfm_cpp`, dfm, W_raw, vocab)
+read_dfm_cpp <- function(dfm, W_raw, vocab, show_progress_bar) {
+    .Call(`_keyATM_read_dfm_cpp`, dfm, W_raw, vocab, show_progress_bar)
 }
 
 #' Run the Collapsed Gibbs sampler for the keyATM Base